Arreglando detalles

+++ src/main/java/com/utn/TPFinal/controller/model/UserController.java
@@ -8,6 +8,7 @@ import com.utn.TPFinal.exceptions.UserNotExistException;
 import com.utn.TPFinal.exceptions.ValidationException;
 import com.utn.TPFinal.model.User;
 import com.utn.TPFinal.service.UserService;
+import org.junit.platform.commons.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.stereotype.Controller;
@@ -23,52 +24,84 @@ public class UserController {
         this.userService = userService;
     }
 
-    public Integer addClient(UserPhoneDto clientPhone) throws JpaSystemException {
-        return userService.addClientPhone(clientPhone);
+    public Integer addClient(UserPhoneDto clientPhone) throws JpaSystemException, ValidationException {
+        if (clientPhone.isValid()) {
+            return userService.addClientPhone(clientPhone);
+        } else {
+            throw new ValidationException("Wrong parameters (empty, null, or wrong)");
+        }
     }
 
-    public void deleteClient(String dni) throws JpaSystemException {
-        userService.deleteClientPhone(dni);
+    public void deleteClient(String dni) throws JpaSystemException, ValidationException {
+        if (!StringUtils.isBlank(dni)) {
+            userService.deleteClientPhone(dni);
+        } else {
+            throw new ValidationException("Wrong parameters (empty, null, or wrong)");
+        }
     }
 
-    public void suspendClient(String dni) throws JpaSystemException {
-        userService.suspendClient(dni);
+    public void suspendClient(String dni) throws JpaSystemException, ValidationException {
+        if (!StringUtils.isBlank(dni)) {
+            userService.suspendClient(dni);
+        } else {
+            throw new ValidationException("Wrong parameters (empty, null, or wrong)");
+        }
     }
 
-    public void modifyClient(UserPhoneModifyDto clientPhone) throws JpaSystemException {
-        userService.modifyClientPhone(clientPhone);
+    public void modifyClient(UserPhoneModifyDto clientPhone) throws JpaSystemException, ValidationException {
+        if (clientPhone.isValid()) {
+            userService.modifyClientPhone(clientPhone);
+        } else {
+            throw new ValidationException("Wrong parameters (empty, null, or wrong)");
+        }
     }
 
     public List<User> getAllEmployee() {
         return userService.getAllEmployee();
     }
 
-    public Integer addEmployee(EmployeeDto employee) throws JpaSystemException {
-        return userService.addEmployee(employee);
+    public Integer addEmployee(EmployeeDto employee) throws JpaSystemException, ValidationException {
+        if (employee.isValid()) {
+            return userService.addEmployee(employee);
+        } else {
+            throw new ValidationException("Wrong parameters (empty, null, or wrong)");
+        }
     }
 
     public List<User> getAllClients() {
         return userService.getAllClient();
     }
 
-    public User getEmployee(String dni) {
-        return userService.getEmployeeDni(dni);
+    public User getEmployee(String dni) throws ValidationException {
+        if (!StringUtils.isBlank(dni)) {
+            return userService.getEmployeeDni(dni);
+        } else {
+            throw new ValidationException("Wrong parameters (empty, null, or wrong)");
+        }
     }
 
     public User login(LoginRequestDto login) throws UserNotExistException, ValidationException {
-        if ((login.getDni() != null) && (login.getPassword() != null)) {
+        if (login.isValid()) {
             return userService.login(login);
         } else {
             throw new ValidationException("username and password must have a value");
         }
     }
 
-    public User getClient(String dni) {
-        return userService.findClientByDni(dni);
+    public User getClient(String dni) throws ValidationException {
+        if (!StringUtils.isBlank(dni)) {
+            return userService.findClientByDni(dni);
+        } else {
+            throw new ValidationException("Wrong parameters (empty, null, or wrong)");
+        }
     }
 
-    public void reactiveClient(String dni) throws JpaSystemException {
-        userService.reactiveClient(dni);
+    public void reactiveClient(String dni) throws JpaSystemException, ValidationException {
+        if (!StringUtils.isBlank(dni)) {
+            userService.reactiveClient(dni);
+        } else {
+            throw new ValidationException("Wrong parameters (empty, null, or wrong)");
+        }
     }
 }
